+++ app/controllers/external_controller.rb
@@ -1,0 +1,8 @@
+# Displays publicly available pages
+class ExternalController < ApplicationController
+  def contact
+  end
+
+  def sites
+  end
+end